backend/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile ./requirements.in
#
appdirs==1.4.4
    # via black
asgiref==3.4.1
    # via uvicorn
black==21.7b0
    # via -r ./requirements.in
click==8.0.1
    # via
    #   black
    #   pip-tools
    #   uvicorn
dnspython==2.1.0
    # via
    #   -r ./requirements.in
    #   email-validator
email-validator==1.1.3
    # via pydantic
fastapi==0.68.1
    # via -r ./requirements.in
flake8==3.9.2
    # via -r ./requirements.in
h11==0.12.0
    # via uvicorn
httptools==0.2.0
    # via uvicorn
idna==3.2
    # via email-validator
mccabe==0.6.1
    # via flake8
motor==2.5.1
    # via -r ./requirements.in
mypy-extensions==0.4.3
    # via black
pathspec==0.9.0
    # via black
pep517==0.11.0
    # via pip-tools
pip-tools==6.2.0
    # via -r ./requirements.in
pycodestyle==2.7.0
    # via flake8
pydantic[email]==1.8.2
    # via
    #   -r ./requirements.in
    #   fastapi
pyflakes==2.3.1
    # via flake8
pymongo==3.12.0
    # via motor
python-dotenv==0.19.0
    # via
    #   -r ./requirements.in
    #   uvicorn
pyyaml==5.4.1
    # via uvicorn
regex==2021.8.21
    # via black
starlette==0.14.2
    # via fastapi
tomli==1.2.1
    # via
    #   black
    #   pep517
typing-extensions==3.10.0.0
    # via pydantic
uvicorn[standard]==0.15.0
    # via -r ./requirements.in
uvloop==0.16.0
    # via uvicorn
watchgod==0.7
    # via uvicorn
websockets==9.1
    # via uvicorn
wheel==0.37.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
